+++ src/auto-reply/reply/abort.ts
@@ -19,8 +19,15 @@ import { parseAgentSessionKey } from "../../routing/session-key.js";
 import { resolveCommandAuthorization } from "../command-auth.js";
 import { normalizeCommandBody, type CommandNormalizeOptions } from "../commands-registry.js";
 import type { FinalizedMsgContext, MsgContext } from "../templating.js";
+import {
+  applyAbortCutoffToSessionEntry,
+  resolveAbortCutoffFromContext,
+  shouldPersistAbortCutoff,
+} from "./abort-cutoff.js";
 import { stripMentions, stripStructuralPrefixes } from "./mentions.js";
 import { clearSessionQueues } from "./queue.js";
+
+export { resolveAbortCutoffFromContext, shouldSkipMessageByAbortCutoff } from "./abort-cutoff.js";
 
 const ABORT_TRIGGERS = new Set([
   "stop",
@@ -110,80 +117,6 @@ export function getAbortMemory(key: string): boolean | undefined {
     return undefined;
   }
   return ABORT_MEMORY.get(normalized);
-}
-
-export type AbortCutoff = {
-  messageSid?: string;
-  timestamp?: number;
-};
-
-export function resolveAbortCutoffFromContext(ctx: MsgContext): AbortCutoff | undefined {
-  const messageSid =
-    (typeof ctx.MessageSidFull === "string" && ctx.MessageSidFull.trim()) ||
-    (typeof ctx.MessageSid === "string" && ctx.MessageSid.trim()) ||
-    undefined;
-  const timestamp =
-    typeof ctx.Timestamp === "number" && Number.isFinite(ctx.Timestamp) ? ctx.Timestamp : undefined;
-  if (!messageSid && timestamp === undefined) {
-    return undefined;
-  }
-  return { messageSid, timestamp };
-}
-
-function toNumericMessageSid(value: string | undefined): bigint | undefined {
-  const trimmed = value?.trim();
-  if (!trimmed || !/^\d+$/.test(trimmed)) {
-    return undefined;
-  }
-  try {
-    return BigInt(trimmed);
-  } catch {
-    return undefined;
-  }
-}
-
-export function shouldSkipMessageByAbortCutoff(params: {
-  cutoffMessageSid?: string;
-  cutoffTimestamp?: number;
-  messageSid?: string;
-  timestamp?: number;
-}): boolean {
-  const cutoffSid = params.cutoffMessageSid?.trim();
-  const currentSid = params.messageSid?.trim();
-  if (cutoffSid && currentSid) {
-    const cutoffNumeric = toNumericMessageSid(cutoffSid);
-    const currentNumeric = toNumericMessageSid(currentSid);
-    if (cutoffNumeric !== undefined && currentNumeric !== undefined) {
-      return currentNumeric <= cutoffNumeric;
-    }
-    if (currentSid === cutoffSid) {
-      return true;
-    }
-  }
-  if (
-    typeof params.cutoffTimestamp === "number" &&
-    Number.isFinite(params.cutoffTimestamp) &&
-    typeof params.timestamp === "number" &&
-    Number.isFinite(params.timestamp)
-  ) {
-    return params.timestamp <= params.cutoffTimestamp;
-  }
-  return false;
-}
-
-function shouldPersistAbortCutoff(params: {
-  commandSessionKey?: string;
-  targetSessionKey?: string;
-}): boolean {
-  const commandSessionKey = params.commandSessionKey?.trim();
-  const targetSessionKey = params.targetSessionKey?.trim();
-  if (!commandSessionKey || !targetSessionKey) {
-    return true;
-  }
-  // Native targeted /stop can run from a slash/session-control key while the
-  // actual target session uses different message id/timestamp spaces.
-  // Persist cutoff only when command source and target are the same session.
-  return commandSessionKey === targetSessionKey;
 }
 
 function pruneAbortMemory(): void {
@@ -379,8 +312,7 @@ export async function tryFastAbortFromMessage(params: {
       : undefined;
     if (entry && key) {
       entry.abortedLastRun = true;
-      entry.abortCutoffMessageSid = abortCutoff?.messageSid;
-      entry.abortCutoffTimestamp = abortCutoff?.timestamp;
+      applyAbortCutoffToSessionEntry(entry, abortCutoff);
       entry.updatedAt = Date.now();
       store[key] = entry;
       await updateSessionStore(storePath, (nextStore) => {
@@ -389,8 +321,7 @@ export async function tryFastAbortFromMessage(params: {
           return;
         }
         nextEntry.abortedLastRun = true;
-        nextEntry.abortCutoffMessageSid = abortCutoff?.messageSid;
-        nextEntry.abortCutoffTimestamp = abortCutoff?.timestamp;
+        applyAbortCutoffToSessionEntry(nextEntry, abortCutoff);
         nextEntry.updatedAt = Date.now();
         nextStore[key] = nextEntry;
       });
